+++ api/tests/school/integration/application/assessment/controller_test.js
@@ -1,5 +1,5 @@
 import { expect, hFake, sinon } from '../../../../test-helper.js';
-import { assessmentController } from '../../../../../src/school/application/assessments/assessment-controller.js';
+import { assessmentController } from '../../../../../src/school/application/assessments/controller.js';
 import { usecases } from '../../../../../src/school/shared/usecases/index.js';
 import { MissionAssessment } from '../../../../../src/school/domain/models/MissionAssessment.js';
